Revert "MAVEN-152 rollback liferay-maven-plugin to 6.2.10.11 codebase"
This reverts commit 79ed32ad4b80fb81f73c82ee47dcd6c405925aef.
