+++ plugins/liferay-maven-plugin/local-build.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+mvn clean install -DskipTests
+
+mvn verify
