+++ plugins/liferay-maven-plugin/pom.xml
@@ -24,6 +24,13 @@
 				</configuration>
 			</plugin>
 		</plugins>
+		<testResources>
+			<testResource>
+				<directory>src/test/resources/projects</directory>
+				<filtering>true</filtering>
+				<targetPath>projects</targetPath>
+			</testResource>
+		</testResources>
 	</build>
 	<dependencies>
 
@@ -36,6 +43,16 @@
 		</dependency>
 		<dependency>
 			<groupId>org.apache.maven</groupId>
+			<artifactId>maven-compat</artifactId>
+			<version>${maven.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>org.apache.maven</groupId>
+			<artifactId>maven-core</artifactId>
+			<version>${maven.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>org.apache.maven</groupId>
 			<artifactId>maven-model</artifactId>
 			<version>${maven.version}</version>
 		</dependency>
@@ -44,23 +61,13 @@
 			<artifactId>maven-plugin-api</artifactId>
 			<version>${maven.version}</version>
 		</dependency>
-		<dependency>
-			<groupId>org.apache.maven</groupId>
-			<artifactId>maven-project</artifactId>
-			<version>${maven.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.maven.shared</groupId>
-			<artifactId>maven-invoker</artifactId>
-			<version>${maven.version}</version>
-		</dependency>
 
 		<!-- External -->
 
 		<dependency>
 			<groupId>commons-io</groupId>
 			<artifactId>commons-io</artifactId>
-			<version>1.4</version>
+			<version>2.4</version>
 		</dependency>
 		<dependency>
 			<groupId>commons-lang</groupId>
@@ -86,6 +93,19 @@
 			<groupId>org.codehaus.plexus</groupId>
 			<artifactId>plexus-io</artifactId>
 			<version>1.0-alpha-1</version>
+		</dependency>
+
+		<!-- Unit testing -->
+
+		<dependency>
+			<groupId>org.apache.maven.plugin-testing</groupId>
+			<artifactId>maven-plugin-testing-harness</artifactId>
+			<version>3.3.0</version>
+		</dependency>
+		<dependency>
+			<groupId>org.apache.maven.shared</groupId>
+			<artifactId>maven-verifier</artifactId>
+			<version>1.5</version>
 		</dependency>
 	</dependencies>
 </project>
