+++ plugins/liferay-maven-plugin/src/main/java/com/liferay/maven/plugins/LangBuilderMojo.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -20,7 +20,7 @@ import java.io.File;
  * @author Mika Koivisto
  * @goal   build-lang
  */
-public class LangBuilderMojo extends AbstractLiferayMojo {
+public class LangBuilderMojo extends AbstractToolsLiferayMojo {
 
 	protected void doExecute() throws Exception {
 		File file = new File(langDir, langFile + ".properties");
