+++ plugins/liferay-maven-plugin/src/main/java/com/liferay/maven/plugins/PluginDeployerMojo.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -18,6 +18,8 @@ import com.liferay.maven.plugins.util.CopyTask;
 
 import java.io.File;
 
+import org.apache.maven.plugin.MojoExecutionException;
+
 /**
  * @author Mika Koivisto
  * @author Thiago Moreira
@@ -25,7 +27,11 @@ import java.io.File;
  */
 public class PluginDeployerMojo extends AbstractLiferayMojo {
 
-	protected void doExecute() throws Exception {
+	public void execute() throws MojoExecutionException {
+		if (!isLiferayProject()) {
+			return;
+		}
+
 		if (warFile.exists()) {
 			getLog().info(
 				"Deploying " + warFileName + " to " +
@@ -35,7 +41,7 @@ public class PluginDeployerMojo extends AbstractLiferayMojo {
 				warFile, autoDeployDir, warFileName, null, true, true);
 		}
 		else {
-			getLog().warn(warFileName + " does not exist");
+			getLog().warn(warFile.getAbsolutePath() + " does not exist");
 		}
 	}
