+++ plugins/liferay-maven-plugin/src/main/java/com/liferay/maven/plugins/PluginDirectDeployerMojo.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -27,7 +27,7 @@ import org.apache.maven.plugin.MojoExecutionException;
  * @author Thiago Moreira
  * @goal   direct-deploy
  */
-public class PluginDirectDeployerMojo extends AbstractLiferayMojo {
+public class PluginDirectDeployerMojo extends AbstractToolsLiferayMojo {
 
 	protected void deployExt() throws Exception {
 		String artifactId = project.getArtifactId();
