+++ plugins/liferay-maven-plugin/src/main/java/com/liferay/maven/plugins/SassToCssBuilderMojo.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -30,7 +30,7 @@ import org.apache.commons.io.filefilter.FileFilterUtils;
  * @goal   build-css
  * @phase  process-sources
  */
-public class SassToCssBuilderMojo extends AbstractLiferayMojo {
+public class SassToCssBuilderMojo extends AbstractToolsLiferayMojo {
 
 	protected void doExecute() throws Exception {
 		FileFilter fileFilter = FileFilterUtils.orFileFilter(
