+++ plugins/liferay-maven-plugin/src/main/java/com/liferay/maven/plugins/ThemeMergeMojo.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -43,7 +43,7 @@ import org.dom4j.Element;
  * @goal   theme-merge
  * @phase  process-sources
  */
-public class ThemeMergeMojo extends AbstractLiferayMojo {
+public class ThemeMergeMojo extends AbstractToolsLiferayMojo {
 
 	protected void cleanUpTemplates(File templatesDir) {
 		File initFile = new File(templatesDir, "init." + themeType);
