+++ plugins/liferay-maven-plugin/src/main/java/com/liferay/maven/plugins/ThumbnailBuilderMojo.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -21,7 +21,7 @@ import java.io.File;
  * @goal   build-thumbnail
  * @phase  process-sources
  */
-public class ThumbnailBuilderMojo extends AbstractLiferayMojo {
+public class ThumbnailBuilderMojo extends AbstractToolsLiferayMojo {
 
 	protected void doExecute() throws Exception {
 		String[] args = new String[5];
