+++ plugins/liferay-maven-plugin/src/main/java/com/liferay/maven/plugins/WSDDBuilderMojo.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License as published by the Free
@@ -20,7 +20,7 @@ import java.io.File;
  * @author Mika Koivisto
  * @goal   build-wsdd
  */
-public class WSDDBuilderMojo extends AbstractLiferayMojo {
+public class WSDDBuilderMojo extends AbstractToolsLiferayMojo {
 
 	protected void doExecute() throws Exception {
 		File serviceFile = new File(serviceFileName);
