+++ plugins/liferay-maven-plugin/src/test/resources/projects/plugindeployer/missingWarFile/pom.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0"?>
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>it</groupId>
+	<artifactId>missingWarFile</artifactId>
+	<packaging>war</packaging>
+	<name>missingWarFile</name>
+	<version>1.0</version>
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>com.liferay.maven.plugins</groupId>
+				<artifactId>liferay-maven-plugin</artifactId>
+				<version>${project.version}</version>
+				<configuration>
+					<autoDeployDir>${project.build.testOutputDirectory}/deploy</autoDeployDir>
+				</configuration>
+			</plugin>
+		</plugins>
+	</build>
+</project>
