+++ plugins/liferay-maven-plugin/src/test/resources/projects/servicebuilder/testProject/pom.xml
@@ -1,0 +1,58 @@
+<?xml version="1.0"?>
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>it</groupId>
+	<artifactId>testProject</artifactId>
+	<packaging>pom</packaging>
+	<name>testProject</name>
+	<version>1.0</version>
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>com.liferay.maven.plugins</groupId>
+				<artifactId>liferay-maven-plugin</artifactId>
+				<version>${liferay.maven.plugin.version}</version>
+				<configuration>
+					<liferayVersion>${liferay.version}</liferayVersion>
+				</configuration>
+			</plugin>
+			<plugin>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>2.5</version>
+				<configuration>
+					<encoding>UTF-8</encoding>
+					<source>1.6</source>
+					<target>1.6</target>
+				</configuration>
+			</plugin>
+			<plugin>
+				<artifactId>maven-resources-plugin</artifactId>
+				<version>2.5</version>
+				<configuration>
+					<encoding>UTF-8</encoding>
+				</configuration>
+			</plugin>
+		</plugins>
+	</build>
+	<modules>
+		<module>testProject-portlet</module>
+		<module>testProject-portlet-service</module>
+	</modules>
+	<profiles>
+		<profile>
+			<id>6.1.2</id>
+			<properties>
+				<liferay.version>6.1.2</liferay.version>
+				<liferay.maven.plugin.version>${project.version}</liferay.maven.plugin.version>
+			</properties>
+		</profile>
+		<profile>
+			<id>6.2.2</id>
+			<properties>
+				<liferay.version>6.2.2</liferay.version>
+				<liferay.maven.plugin.version>${project.version}</liferay.maven.plugin.version>
+			</properties>
+		</profile>
+	</profiles>
+</project>
