+++ plugins/liferay-maven-plugin/src/test/resources/projects/servicebuilder/testProject/testProject-portlet-service/pom.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0"?>
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+	<parent>
+		<groupId>it</groupId>
+		<artifactId>testProject</artifactId>
+		<version>1.0</version>
+		<relativePath>../pom.xml</relativePath>
+	</parent>
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>it</groupId>
+	<artifactId>testProject-portlet-service</artifactId>
+	<packaging>jar</packaging>
+	<name>testProject Portlet Service</name>
+	<version>1.0</version>
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>com.liferay.maven.plugins</groupId>
+				<artifactId>liferay-maven-plugin</artifactId>
+				<version>${liferay.maven.plugin.version}</version>
+				<configuration>
+					<webappBaseDir>${project.build.testOutputDirectory}/projects/servicebuilder/testProject/testProject-portlet</webappBaseDir>
+				</configuration>
+			</plugin>
+		</plugins>
+	</build>
+	<dependencies>
+		<dependency>
+			<groupId>com.liferay.portal</groupId>
+			<artifactId>portal-service</artifactId>
+			<version>${liferay.version}</version>
+			<scope>provided</scope>
+		</dependency>
+	</dependencies>
+</project>
