+++ plugins/liferay-maven-plugin/src/test/resources/unit/poms/plugin-deployer-mojo-pom.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0"?>
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>com.liferay.maven.plugins</groupId>
+	<artifactId>plugin-deployer-mojo-tests</artifactId>
+	<packaging>war</packaging>
+	<name>Plugin Deployer Mojo Tests</name>
+	<version>1.0.0-SNAPSHOT</version>
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>com.liferay.maven.plugins</groupId>
+				<artifactId>liferay-maven-plugin</artifactId>
+				<version>${project.version}</version>
+				<configuration>
+					<autoDeployDir>target/deploy</autoDeployDir>
+					<liferayVersion>7.0.0</liferayVersion>
+					<warFile>src/test/resources/unit/wars/empty.war</warFile>
+					<warFileName>empty.war</warFileName>
+					<project implementation="org.apache.maven.plugin.testing.stubs.MavenProjectStub">
+						<packaging>war</packaging>
+					</project>
+				</configuration>
+			</plugin>
+		</plugins>
+	</build>
+</project>
